Create FaultyGuiMenu
The options work but I couldn't get the buttons to display. Also the options page was the same screen as the main page so I got confused on how to get the buttons to display/exist on the option page and not the menu. No way to test until we can return to previous page.